python-app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.70
botocore==1.35.70
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.10.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
llamaapi==0.1.36
MarkupSafe==3.0.2
minio==7.2.12
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.55.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
prft==1.3.0
propcache==0.2.0
protobuf==5.29.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.1.3
yarl==1.18.0
